+++ warden/x/warden/keeper/msg_server_update_keychain.go
@@ -21,7 +21,6 @@ func (k msgServer) UpdateKeychain(goCtx context.Context, msg *types.MsgUpdateKey
 		return nil, fmt.Errorf("keychain updates should be requested by admins")
 	}
 
-	kr.SetStatus(msg.IsActive)
 	kr.SetFees(msg.KeychainFees)
 	if msg.Description != "" {
 		kr.SetDescription(msg.Description)